% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagmatrix.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Create lead/lags of a variable}
\usage{
lagmatrix(x, lag)
}
\arguments{
\item{x}{input variable.}

\item{lag}{vector of leads and lags. Positive numbers are lags, negative are leads. O is the original \code{x}.}
}
\value{
An array with the resulting leads and lags (columns).
}
\description{
Create an array with lead/lags of an input variable.
}
\examples{
  x <- rnorm(10)
  lagmatrix(x,c(0,1,-1))

}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
