% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leadtrail.R
\name{leadtrail}
\alias{leadtrail}
\title{Remove leading/training zeros/NAs}
\usage{
leadtrail(x, rm = c("zeros", "na"), lead = c(TRUE, FALSE),
  trail = c(TRUE, FALSE))
}
\arguments{
\item{x}{vector of values to check.}

\item{rm}{what to remove, can be \code{"zeros"} or \code{"na"}.}

\item{lead}{If \code{TRUE}, then leading values are removed.}

\item{trail}{If \code{TRUE}, then trailing values are removed.}
}
\value{
Resulting vector.
}
\description{
Remove leading or trailing zeros or NAs from a vector.
}
\examples{
x <- c(rep(0,5),rnorm(100),rep(0,5))
leadtrail(x)

}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}.
}
