% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{referrals}
\alias{referrals}
\title{NHS A&E Referrals}
\description{
Monthly Accident & Emergency referrals for England and Wales.
}
\references{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/ae-waiting-times-and-activity/}
}
\keyword{data}
