% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sthief.R
\name{Sthief}
\alias{Sthief}
\title{Temporal hierarchy S matrix}
\usage{
Sthief(y)
}
\arguments{
\item{y}{input time series (a \code{ts} object) or an integer.}
}
\value{
S matrix.
}
\description{
Calculate the temporal hierarchy summing matrix S for a given time series of seasonal periodicity.
}
\examples{
Sthief(AirPassengers)

}
\references{
Athanasopoulos, G., Hyndman, R. J., Kourentzes, N., & Petropoulos, F. (2017). \href{https://kourentzes.com/forecasting/2017/02/27/forecasting-with-temporal-hierarchies-3/}{Forecasting with temporal hierarchies}. European Journal of Operational Research, 262(1), 60-74.
}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}.
}
\keyword{ts}
