% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsutils-package.R
\docType{package}
\name{tsutils-package}
\alias{tsutils-package}
\alias{_PACKAGE}
\title{tsutils: Time Series Exploration, Modelling and Forecasting}
\description{
Includes: (i) tests and visualisations that can help the modeller explore time series components and perform decomposition; (ii) modelling shortcuts, such as functions to construct lagmatrices and seasonal dummy variables of various forms; (iii) an implementation of the Theta method; (iv) tools to facilitate the design of the forecasting process, such as ABC-XYZ analyses; and (v) "quality of life" functions, such as treating time series for trailing and leading values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trnnick/tsutils/}
}

}
\author{
\strong{Maintainer}: Nikolaos Kourentzes \email{nikolaos@kourentzes.com}

Other contributors:
\itemize{
  \item Ivan Svetunkov \email{ivan@svetunkov.ru} [contributor]
  \item Oliver Schaer \email{info@oliverschaer.ch} [contributor]
}

}
\keyword{internal}
